Decoupling and small fix
