+++ src/Components/Endpoints/src/DependencyInjection/CookieTempDataProvider.cs
@@ -17,7 +17,8 @@ internal sealed partial class CookieTempDataProvider : ITempDataProvider
     private const int MaxEncodedLength = 4050;
     private readonly IDataProtector _dataProtector;
 
-    public CookieTempDataProvider(IDataProtectionProvider dataProtectionProvider)
+    public CookieTempDataProvider(
+        IDataProtectionProvider dataProtectionProvider)
     {
         _dataProtector = dataProtectionProvider.CreateProtector(Purpose);
     }
@@ -45,7 +46,7 @@ internal sealed partial class CookieTempDataProvider : ITempDataProvider
             var convertedData = new Dictionary<string, object?>();
             foreach (var kvp in dataFromCookie)
             {
-                convertedData[kvp.Key] = ConvertJsonElement(kvp.Value);
+                convertedData[kvp.Key] = TempDataSerializer.ConvertJsonElement(kvp.Value);
             }
             return convertedData;
         }
@@ -70,7 +71,7 @@ internal sealed partial class CookieTempDataProvider : ITempDataProvider
     {
         foreach (var kvp in values)
         {
-            if (!CanSerializeType(kvp.Value?.GetType() ?? typeof(object)))
+            if (!TempDataSerializer.CanSerializeType(kvp.Value?.GetType() ?? typeof(object)))
             {
                 throw new InvalidOperationException($"TempData cannot store values of type '{kvp.Value?.GetType()}'.");
             }
@@ -111,89 +112,6 @@ internal sealed partial class CookieTempDataProvider : ITempDataProvider
             Secure = context.Request.IsHttps,
             Path = context.Request.PathBase.HasValue ? context.Request.PathBase.Value : "/",
         });
-    }
-
-    private static object? ConvertJsonElement(JsonElement element)
-    {
-        switch (element.ValueKind)
-        {
-            case JsonValueKind.String:
-                if (element.TryGetGuid(out var guid))
-                {
-                    return guid;
-                }
-                if (element.TryGetDateTime(out var dateTime))
-                {
-                    return dateTime;
-                }
-                return element.GetString();
-            case JsonValueKind.Number:
-                return element.GetInt32();
-            case JsonValueKind.True:
-            case JsonValueKind.False:
-                return element.GetBoolean();
-            case JsonValueKind.Null:
-                return null;
-            case JsonValueKind.Array:
-                return DeserializeArray(element);
-            case JsonValueKind.Object:
-                return DeserializeDictionaryEntry(element);
-            default:
-                throw new InvalidOperationException($"TempData cannot deserialize value of type '{element.ValueKind}'.");
-        }
-    }
-
-    private static object? DeserializeArray(JsonElement arrayElement)
-    {
-        var arrayLength = arrayElement.GetArrayLength();
-        if (arrayLength == 0)
-        {
-            return null;
-        }
-        if (arrayElement[0].ValueKind == JsonValueKind.String)
-        {
-            var array = new List<string?>(arrayLength);
-            foreach (var item in arrayElement.EnumerateArray())
-            {
-                array.Add(item.GetString());
-            }
-            return array.ToArray();
-        }
-        else if (arrayElement[0].ValueKind == JsonValueKind.Number)
-        {
-            var array = new List<int>(arrayLength);
-            foreach (var item in arrayElement.EnumerateArray())
-            {
-                array.Add(item.GetInt32());
-            }
-            return array.ToArray();
-        }
-        throw new InvalidOperationException($"TempData cannot deserialize array of type '{arrayElement[0].ValueKind}'.");
-    }
-
-    private static Dictionary<string, string?> DeserializeDictionaryEntry(JsonElement objectElement)
-    {
-        var dictionary = new Dictionary<string, string?>(StringComparer.Ordinal);
-        foreach (var item in objectElement.EnumerateObject())
-        {
-            dictionary[item.Name] = item.Value.GetString();
-        }
-        return dictionary;
-    }
-
-    private static bool CanSerializeType(Type type)
-    {
-        type = Nullable.GetUnderlyingType(type) ?? type;
-        return
-            type.IsEnum ||
-            type == typeof(int) ||
-            type == typeof(string) ||
-            type == typeof(bool) ||
-            type == typeof(DateTime) ||
-            type == typeof(Guid) ||
-            typeof(ICollection<int>).IsAssignableFrom(type) ||
-            typeof(ICollection<string>).IsAssignableFrom(type) ||
-            typeof(IDictionary<string, string>).IsAssignableFrom(type);
     }
 
     private static partial class Log
